[img-editor] Added dedicated brush size control.

+++ Telegram/SourceFiles/editor/photo_editor_inner_common.h
@@ -26,7 +26,7 @@ struct PhotoEditorMode {
 };
 
 struct Brush {
-	float sizeRatio = 0.;
+	float64 sizeRatio = 0.;
 	QColor color;
 };
